depend on r7rs 0.0.5 and r6rs-bytevectors 1.0.2
There was a bug in r7rs 0.0.4 and earlier whereby r6rs-bytevectors
redefined the bytevector-copy! from r7rs even though it didn't import
it.  That procedure has a different call signature in r6rs than it does
in r7rs, so the bug resulted in the wrong one being visible when
bytestructures was imported and r7rs was also imported.  r7rs 0.0.5
fixed that issue.